pyproject.toml
[tool.poetry]
name = "products-backend"
version = "0.1.0"
description = ""
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "^5.0"
djangorestframework = "^3.15"
django-filter = "^24.2"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.4"
django-environ = "^0.11"
dj-database-url = "^2.2"
psycopg = {extras = ["binary"], version = "^3.2.10"}
whitenoise = "^6.7"
django-extensions = "^3.2"
djangorestframework-simplejwt = {version = "^5.3", extras = ["crypto"]}
django-guardian = "^2.4"
django-redis = "^5.4"
redis = "^5.0"
django-rq = "^2.10"
rq = "^1.16"
django-storages = "^1.14"
boto3 = "^1.34"
django-health-check = "^3.18"
django-prometheus = "^2.3"
sentry-sdk = {version = "^2.15", extras = ["django"]}
drf-nested-routers = "^0.93.5"
drf-rw-serializers = "^1.1.0"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "model-bakery (>=1.20.5,<2.0.0)",
    "coverage[toml] (>=7.10.7,<8.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "ruff (>=0.13.1,<0.14.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "types-python-dateutil (>=2.9.0.20250822,<3.0.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "types-psycopg2 (>=2.9.21.20250915,<3.0.0.0)"
]
